fix(webpack): make ChunkNames compatible with external dll-delegated chunk
The modules in chunk.modules[0] does not necessarily be a NormalModule which has well-defined userRequest and rawRequest. It could be a DelegatedModule when the chunk is sealed in an external dll files. In this case, the rawRequest will be undefined since a DelegatedModule does not have this property. However, the userRequest property can supplement rawRequest well in this situation.